Add test of max num among siblings (#17)
